+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Please refer to :doc:`release notes<release_notes>`.
 4.0.0-alpha.6 (unreleased)
 --------------------------
 
+- Added debug for missing fields & attributes during import (#1571)
 - added :meth:`~import_export.widgets.ForeignKeyWidget.get_lookup_kwargs` to make it easier to override object
   lookup (#1651)
 - removed unused variable ``Result.new_record`` (#1640)